LPS-123796 Move some Actions logic from the EditorVariant to the Actions component itself
